Day 16 P2: Maze of valves, now with elephants!
This was the hardest programming exercise I have ever done. University
labs and exams doesn't even come close. I need way more practice with
memoization and saving states.

I spent several hours trying to optimize my solution to get
the time and memory usage down. I got it several times faster but it was
still too slow to finish in a reasonable time. So I just started pruning
states that had low total pressure since they were unlikely to be the
best ones. I ended up cutting the bottom 90% of all states and the
program finished in about a minute. I was very surprised that the
solution was correct since it's definitely not guaranteed...

I really want to solve this in a good way to learn how to do it, but
that has to wait until AoC is over...