Day 17: Scuffed Tetris
Part 1 was fairly straight forward. I had a really stupid bug with a
newline character in my jet string that cost like 30 minutes of
debugging...

Part 2 got me sighing initially, then I realized that the only way to
actually solve this is if there's a repeating pattern somewhere. So I
dumped out a large part of the grid looked at it manually to verify that
there was a repeating pattern. Then I let it run for a bit, took a
snapshot of the latest 20ish lines, let it keep running until those
lines repeated, then did some fancy math to skip a bunch of iterations,
then let it simulate the last couple of iterations manually again.

Was a pretty fun challenge, especially when compared to yesterday!