Day 21: The return of mathematically adept monkeys
Another easy one! I wrote part 2 to return the equation with "humn" as
"x" and then solving it, but it was too huge for Wolfram Alpha, and
ChatGPT can't do math for shit apparently. So I just started testing
different numbers and arrived at the correct one in like 5 minutes. Way
easier!