Day 22 P2: Crappy Cubes
This was a horrible hardcoded solution. Basically I hardcoded the
relations between the faces on the example cube, then realized that the
real cube has completely different relations... Had to go to work and
when I got home I hardcoded the real cube as well, realizing that it
would be really hard to debug this if something was wrong since it
wouldn't work on the example. Luckily the Almighty Stormfather himself
came down and blessed me, making me write perfect code that got the
correct answer on the first try! Thanks!