Day 23: Social distancing
This was a fun one! I really need to figure out how to work with
expanding grids though, here I just padded the grid with 200 empty tiles
in every direction but it makes the computations take a long time. Would
be great if the grid could automatically expand somehow, will have to
work on that!